Move safeJsonParse out of util.ts so that it can be used during initialization
